# src/main/resources/static/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DMC API",
    "description": "Move your app forward with the DMC API",
    "version": "1.0.0"
  },
  "host": "api.dmc.com",
  "schemes": [
    "https"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/login": {
      "post": {
        "summary": "登录",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "用户名和密码登录",
            "required": true,
            "schema": {
              "required": [
                "username",
                "password"
              ],
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "description": "用户名",
                  "example": "admin1"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "description": "密码",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "登录结果",
            "schema": {
              "$ref": "#/definitions/SessionInfo"
            }
          }
        }
      }
    },

    "/user/editCurrentUserPwd": {
      "post": {
        "summary": "修改密码",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "老密码和新密码",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "type": "string",
                  "description": "老密码"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "description": "密码"
                }
              }
            }
          }
        ],
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
          }
        }
      }
    },
    "/user": {
      "post": {
        "summary": "添加用户",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "description": "用户名"
                },
                "name": {
                  "type": "string",
                  "description": "昵称"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "description": "密码"
                }
              }
            }
          }
        ],
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
          }
        }
      },
      "put": {
        "summary": "更新用户",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int32",
                  "description": "ID"
                },
                "username": {
                  "type": "string",
                  "description": "用户名"
                },
                "name": {
                  "type": "string",
                  "description": "昵称"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "description": "密码"
                }
              }
            }
          }
        ],
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
          }
        }
      }
    },
    "/user/{userId}": {
      "delete": {
        "summary": "删除用户",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],

        "responses": {

        }
      }
    },
    "/resource/menus": {
      "post": {
        "summary": "获取菜单",
        "tags": [
          "Resource"
        ],
        "responses": {
          "200": {
            "description": "资源",
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          }
        }
      }
    },
    "/resource": {
      "post": {
        "summary": "添加资源",
        "parameters": [
          {
            "name": "resource",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "资源名"
                },
                "type": {
                  "type": "number",
                  "description": "0:资源，1:某个操作"
                },
                "pid": {
                  "type": "number",
                  "description": "上级资源ID"
                },
                "url": {
                  "type": "string",
                  "description": "路由路径或者接口地址"
                }
              }
            }
          }
        ],
        "tags": [
          "Resource"
        ],
        "responses": {
          "200": {
            "description": "资源",
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          }
        }
      },
      "put": {
        "summary": "更新资源",
        "parameters": [
          {
            "name": "resource",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64",
                  "description": "资源ID"
                },
                "name": {
                  "type": "string",
                  "description": "资源名"
                },
                "type": {
                  "type": "number",
                  "description": "0:资源，1:某个操作"
                },
                "pid": {
                  "type": "number",
                  "description": "上级资源ID"
                },
                "url": {
                  "type": "string",
                  "description": "路由路径或者接口地址"
                }
              }
            }
          }
        ],
        "tags": [
          "Resource"
        ],
        "responses": {
          "200": {
            "description": "资源",
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          }
        }
      }
    },
    "/resource/{resourceId}": {
      "delete": {
        "tags": [
          "Resource"
        ],
        "summary": "删除资源",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ]
      },
      "get": {
        "tags": [
          "Resource"
        ],
        "summary": "资源详情",
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "唯一ID"
        },
        "name": {
          "type": "string",
          "description": "用户的昵称"
        },
        "username": {
          "type": "string",
          "description": "用户名"
        },
        "password": {
          "type": "string",
          "format": "password",
          "description": "密码"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "modifyTime": {
          "type": "string",
          "format": "date-time",
          "description": "修改时间"
        },
        "roleIds": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "拥有的权限"
        }
      }
    },
    "SessionInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "唯一ID",
          "example": "1"
        },
        "name": {
          "type": "string",
          "description": "用户的昵称",
          "example": "BOSS"
        },
        "resourceList": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "可以访问的资源",
          "example": [
            "/resource/add",
            "/resource/edit",
            "/resource/delete"
          ]
        }
      }
    },
    "Resource": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "资源id"
        },
        "name": {
          "type": "string",
          "description": "资源名"
        },
        "type": {
          "type": "number",
          "description": "0:菜单，1:某个操作"
        },
        "pid": {
          "type": "number",
          "description": "上级资源ID"
        },
        "url": {
          "type": "string",
          "description": "路由路径或者接口地址"
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Resource"
          },
          "description": "下级资源"
        }
      }
    }
  }
}